Enable tests for PeerConnection.SetConfiguration
The tests are divided into two files: one should work with both Go and
the WASM bindings, and one should work with the Go implementation only.
The tests currently fail with the WASM bindings due to the errors not
lining up exactly, but nonetheless this is a good starting point.